requirements.txt
kagglehub
lightning
numpy
sentencepiece
pandas
python-dotenv
torch
torchmetrics
wandb